Moved spending_over_time to its own file
