+++ usaspending_api/search/v2/urls_search.py
@@ -1,9 +1,10 @@
 from django.conf.urls import url
 from usaspending_api.search.v2.views import search
 from usaspending_api.search.v2.views.spending_by_category import SpendingByCategoryVisualizationViewSet
+from usaspending_api.search.v2.views.spending_over_time import SpendingOverTimeVisualizationViewSet
 
 urlpatterns = [
-    url(r'^spending_over_time', search.SpendingOverTimeVisualizationViewSet.as_view()),
+    url(r'^spending_over_time', SpendingOverTimeVisualizationViewSet.as_view()),
     url(r'^spending_by_category', SpendingByCategoryVisualizationViewSet.as_view()),
     url(r'^spending_by_geography', search.SpendingByGeographyVisualizationViewSet.as_view()),
     url(r'^spending_by_award_count', search.SpendingByAwardCountVisualizationViewSet.as_view()),
